+++ Ranger/Services/SharkScope/PlayerSummary/PlayerResponse/GraphData.swift
@@ -15,6 +15,7 @@ struct GraphData: Decodable, Equatable
     
     let id: String
     let dataPoints: [DataPoint]
+    let trendLine: TrendLine
     
     
     struct DataPoint: Decodable, Equatable
@@ -25,6 +26,17 @@ struct GraphData: Decodable, Equatable
         let y: Double
     }
     
+    
+    struct TrendLine: Decodable, Equatable
+    {
+        
+        
+        let slope: Double
+        let offset: Double
+        let max: Double
+    }
+    
+    
     init(from statisticalDataSet: Statistics.StatisticalDataSet?)
     {
         // Only if any.
@@ -34,6 +46,7 @@ struct GraphData: Decodable, Equatable
             // Initialize empty if none.
             self.id = ""
             self.dataPoints = []
+            self.trendLine = TrendLine(slope: 0, offset: 0, max: 0)
             return
         }
         
@@ -41,16 +54,22 @@ struct GraphData: Decodable, Equatable
         self.id = statisticalDataSet.id
         if let data: [Statistics.StatisticalDataSet.Data] = statisticalDataSet.Data
         {
-            self.dataPoints = data.map
+            self.dataPoints = data.enumerated().map
             {
-                eachData in
-                DataPoint(
-                    x: Double(eachData.x) ?? 0.0,
+                (arguments) -> GraphData.DataPoint in
+                let (eachIndex, eachData) = arguments
+                
+                return DataPoint(
+                    x: Double(eachIndex), // Double(eachData.x) ?? 0.0,
                     y: Double(eachData.Y.first?.value ?? "0.0") ?? 0.0
                 )
             }
         }
         else
         { self.dataPoints = [] }
+        
+        // Trendline.
+        let linearRegression = LinearRegression(x: dataPoints.map{$0.x}, y: dataPoints.map{$0.y})
+        self.trendLine = TrendLine(slope: linearRegression.slope, offset: linearRegression.offset, max: linearRegression.max)
     }
 }